requirements.txt
tornado
python-socketio
pyyaml